Add files via upload
This project simulates robot navigation on a 2D grid using Best-First Search and A* algorithms. Users input the grid size, start and goal positions, and obstacles. The program uses Manhattan distance as the heuristic and priority queues to find a path. Best-First is faster but not always optimal, while A* guarantees the shortest path.